refactoring to split out parsing helpers into their own files created tools.js for general-purpose functions
