--- src/components/psshLookup.json
@@ -1,5 +1,0 @@
-{
-    "10 77 ef ec c0 b2 4d 02 ac e3 3c 1e 52 e2 fb 4b": "Clearkey",
-    "9a 04 f0 79 98 40 42 86 ab 92 e6 5b e0 88 5f 95": "PlayReady",
-    "ed ef 8b a9 79 d6 4a ce a3 c8 27 dc d5 1d 21 ed": "WideVine"
-}
